```text chore: Report coverage summary in test.sh script
Update the test.sh script to include a new step for reporting the coverage summary. This step is added after generating the HTML coverage report. By running the command "coverage report -m", the coverage summary will be displayed. This change improves the testing process by providing a summary of the code coverage.